+++ pycantonese/__init__.py
@@ -8,6 +8,7 @@ from pycantonese.jyutping.characters import (
 from pycantonese.jyutping.parse_jyutping import parse_jyutping
 from pycantonese.jyutping.tipa import jyutping_to_tipa, jyutping2tipa
 from pycantonese.jyutping.yale import jyutping_to_yale, jyutping2yale
+from pycantonese.pos_tagging.tagger import pos_tag
 from pycantonese.stop_words import stop_words
 from pycantonese.word_segmentation import segment
 
@@ -24,6 +25,7 @@ __all__ = [
     "jyutping2tipa",
     "jyutping2yale",
     "parse_jyutping",
+    "pos_tag",
     "read_chat",
     "stop_words",
     "segment",