+++ pycantonese/pos_tagging/__init__.py
@@ -1,4 +1,5 @@
 from pycantonese.pos_tagging.hkcancor_to_ud import hkcancor_to_ud
+from pycantonese.pos_tagging.tagger import POSTagger
 
 
-__all__ = ["hkcancor_to_ud"]
+__all__ = ["POSTagger", "hkcancor_to_ud"]