+++ pycantonese/tests/test_pos_tagging/test_hkcancor_to_ud.py
@@ -1,6 +1,6 @@
 import pytest
 
-from pycantonese.pos_tagging.hkcancor_to_ud import hkcancor_to_ud
+from pycantonese.pos_tagging.hkcancor_to_ud import hkcancor_to_ud, _MAP
 
 
 @pytest.mark.parametrize(
@@ -15,3 +15,7 @@ from pycantonese.pos_tagging.hkcancor_to_ud import hkcancor_to_ud
 )
 def test_hkcancor_to_ud(tag, expected):
     assert hkcancor_to_ud(tag) == expected
+
+
+def test_hkcancor_to_ud_all_tags():
+    assert hkcancor_to_ud() == _MAP